# server/plugins/ansible/client/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v1",
    "title" : "Concord Ansible API"
  },
  "host" : "localhost:8001",
  "basePath" : "/",
  "tags" : [ {
    "name" : "Ansible Process"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/api/v1/process/{processInstanceId}/ansible/events" : {
      "get" : {
        "tags" : [ "Ansible Process" ],
        "summary" : "List Ansible events of a specific process",
        "description" : "",
        "operationId" : "listEvents",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "processInstanceId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "host",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "hostGroup",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "playbookId",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AnsibleProcessEventEntry"
              }
            }
          }
        },
        "security" : [ {
          "api_key" : [ ]
        }, {
          "session_key" : [ ]
        }, {
          "ldap" : [ ]
        } ]
      }
    },
    "/api/v1/process/{processInstanceId}/ansible/hosts" : {
      "get" : {
        "tags" : [ "Ansible Process" ],
        "summary" : "List Ansible hosts of a specific process",
        "description" : "",
        "operationId" : "list",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "processInstanceId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "host",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "hostGroup",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "enum" : [ "RUNNING", "CHANGED", "FAILED", "OK", "SKIPPED", "UNREACHABLE" ]
        }, {
          "name" : "statuses",
          "in" : "query",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string",
            "enum" : [ "RUNNING", "CHANGED", "FAILED", "OK", "SKIPPED", "UNREACHABLE" ]
          },
          "collectionFormat" : "multi"
        }, {
          "name" : "playbookId",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "default" : 30,
          "format" : "int32"
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "default" : 0,
          "format" : "int32"
        }, {
          "name" : "sortField",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "enum" : [ "HOST", "DURATION", "STATUS", "HOST_GROUP" ]
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "enum" : [ "ASC", "DESC" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AnsibleHostListResponse"
            }
          }
        },
        "security" : [ {
          "api_key" : [ ]
        }, {
          "session_key" : [ ]
        }, {
          "ldap" : [ ]
        } ]
      }
    },
    "/api/v1/process/{processInstanceId}/ansible/playbooks" : {
      "get" : {
        "tags" : [ "Ansible Process" ],
        "summary" : "List Ansible playbooks of a specific process",
        "description" : "",
        "operationId" : "listPlaybooks",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "processInstanceId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PlaybookEntry"
              }
            }
          }
        },
        "security" : [ {
          "api_key" : [ ]
        }, {
          "session_key" : [ ]
        }, {
          "ldap" : [ ]
        } ]
      }
    },
    "/api/v1/process/{processInstanceId}/ansible/tasks" : {
      "get" : {
        "tags" : [ "Ansible Process" ],
        "summary" : "List Ansible plays of a specific process",
        "description" : "",
        "operationId" : "listTasks",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "processInstanceId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "playId",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/TaskInfo"
              }
            }
          }
        },
        "security" : [ {
          "api_key" : [ ]
        }, {
          "session_key" : [ ]
        }, {
          "ldap" : [ ]
        } ]
      }
    },
    "/api/v1/process/{processInstanceId}/ansible/{playbookId}/plays" : {
      "get" : {
        "tags" : [ "Ansible Process" ],
        "summary" : "List Ansible plays of a specific process",
        "description" : "",
        "operationId" : "listPlays",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "processInstanceId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "playbookId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PlayInfo"
              }
            }
          }
        },
        "security" : [ {
          "api_key" : [ ]
        }, {
          "session_key" : [ ]
        }, {
          "ldap" : [ ]
        } ]
      }
    }
  },
  "securityDefinitions" : {
    "api_key" : {
      "description" : "API key",
      "type" : "apiKey",
      "name" : "api_key",
      "in" : "header"
    },
    "ldap" : {
      "description" : "AD/LDAP username/password",
      "type" : "basic"
    },
    "session_key" : {
      "description" : "Process session key",
      "type" : "apiKey",
      "name" : "session_key",
      "in" : "header"
    }
  },
  "definitions" : {
    "AnsibleHostEntry" : {
      "type" : "object",
      "properties" : {
        "host" : {
          "type" : "string",
          "readOnly" : true
        },
        "hostGroup" : {
          "type" : "string",
          "readOnly" : true
        },
        "status" : {
          "type" : "string",
          "readOnly" : true,
          "enum" : [ "RUNNING", "CHANGED", "FAILED", "OK", "SKIPPED", "UNREACHABLE" ]
        },
        "duration" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        }
      }
    },
    "AnsibleHostListResponse" : {
      "type" : "object",
      "properties" : {
        "hostGroups" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        },
        "items" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/AnsibleHostEntry"
          }
        }
      }
    },
    "AnsibleProcessEventEntry" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "readOnly" : true
        },
        "eventType" : {
          "type" : "string",
          "readOnly" : true
        },
        "data" : {
          "type" : "object",
          "readOnly" : true
        },
        "eventDate" : {
          "type" : "string",
          "format" : "date-time",
          "readOnly" : true
        }
      }
    },
    "PlayInfo" : {
      "type" : "object",
      "properties" : {
        "playId" : {
          "type" : "string",
          "format" : "uuid",
          "readOnly" : true
        },
        "playName" : {
          "type" : "string",
          "readOnly" : true
        },
        "playOrder" : {
          "type" : "integer",
          "format" : "int32",
          "readOnly" : true
        },
        "hostCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "taskCount" : {
          "type" : "integer",
          "format" : "int32",
          "readOnly" : true
        },
        "taskStats" : {
          "type" : "object",
          "readOnly" : true,
          "additionalProperties" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "finishedTaskCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        }
      }
    },
    "PlaybookEntry" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "readOnly" : true
        },
        "name" : {
          "type" : "string",
          "readOnly" : true
        },
        "startedAt" : {
          "type" : "string",
          "format" : "date-time",
          "readOnly" : true
        },
        "hostsCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "failedHostsCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "playsCount" : {
          "type" : "integer",
          "format" : "int32",
          "readOnly" : true
        },
        "failedTasksCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "progress" : {
          "type" : "integer",
          "format" : "int32",
          "readOnly" : true
        },
        "status" : {
          "type" : "string",
          "readOnly" : true,
          "enum" : [ "RUNNING", "OK", "FAILED" ]
        },
        "retryNum" : {
          "type" : "integer",
          "format" : "int32",
          "readOnly" : true
        }
      }
    },
    "TaskInfo" : {
      "type" : "object",
      "properties" : {
        "taskName" : {
          "type" : "string",
          "readOnly" : true
        },
        "type" : {
          "type" : "string",
          "readOnly" : true,
          "enum" : [ "TASK", "SETUP", "HANDLER" ]
        },
        "taskOrder" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "okCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "failedCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "unreachableCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "skippedCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        },
        "runningCount" : {
          "type" : "integer",
          "format" : "int64",
          "readOnly" : true
        }
      }
    }
  }
}